Create Assembly Text Based Game
I was assigned to make a game where it leads to various different paths and it has to be able to go back and forth between the paths I made depending on the story. It has around 36 boxes; each box means a new point and it leads to a different route in the story. The code included variables that you can use in the game that will benefit the user and help them win the game. I added prompts where the user has to hit a certain key to enter a certain path and if they don't hit the key they will be prompt to hit it.